Update Convolution.cl
float

+++ module11/Convolution.cl
@@ -14,7 +14,7 @@
 
 __kernel void convolve(
 	const __global  uint * const input,
-    __constant uint * const mask,
+    __constant float * const mask,
     __global  uint * const output,
     const int inputWidth,
     const int maskWidth)
@@ -22,7 +22,7 @@ __kernel void convolve(
     const int x = get_global_id(0);
     const int y = get_global_id(1);
 
-    uint sum = 0;
+    float sum = 0;
     for (int r = 0; r < maskWidth; r++)
     {
         const int idxIntmp = (y + r) * inputWidth + x;
